+++ python/ray/tune/search/__init__.py
@@ -1,0 +1,4 @@
+from ray.tune.search.search import SearchAlgorithm
+from ray.tune.search.hyperopt import HyperOptAlgorithm
+
+__all__ = ["SearchAlgorithm, HyperOptAlgorithm"]
