+++ python/ray/tune/search/hyperopt.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-
 import numpy as np
 import copy
 try:
@@ -11,28 +10,7 @@ except Exception as e:
     hpo = None
 
 from ray.tune.error import TuneError
-
-
-class SearchAlgorithm():
-    """This class is unaware of Tune trials."""
-    NOT_READY = "NOT_READY"
-
-    def try_suggest():
-        """
-        Returns:
-            Configuration for a trial
-            Suggestion ID
-        """
-        return {}, None
-
-    def on_trial_result():
-        pass
-
-    def on_trial_error():
-        pass
-
-    def on_trial_complete():
-        pass
+from ray.tune.search import SearchAlgorithm
 
 
 class HyperOptAlgorithm(SearchAlgorithm):
